Compatibility Update + Format
Hey alkorin, I made your script compatible with both python 2 and python 3 (I think. Run this script with a test account first just in case, run in both Python2 and 3).

I also added to forms of preemptive user configuration, as well as removed most human error out of ignorance (starting the username with u/ or /u/ or their capital variants won't fuck things up. This is the same with the Subreddit, using the /r/variants are no longer a , and I made the limit for getting the posts 25 by default, just in case a user makes a mistake. The user can set other options.


I also made it PEP257 Compliant.

Part of this was removing the `saved` empty list and `saved_dict` empty dictionary, as in reality you didn't use them.

Before merging, test it with an alt account on both python 2 and 3. I still will later add the private Subreddit enforcement and OAuth compatibility